fix: Ch5.1 - fixes OPTIONS HTTP name and word highlights with errors
Line 203 - Replace the word 'headers' with 'methods' where the text is
referring to methods, not headers.

Line 283 - Correct the highlighting in the expression 'Range Requests'
because the 's' was outside the highlight markup.

Line 289 - Correct the name of the HTTP OPTIONS method.

Line 351 - Format the highlight of the topic, as '>' was outside the
highlight markup.