fix: Ch06.02 - fixes example outputs, semantic error and code block example.
Lines 384, 399, 410 - Correct the property name in the class method and
the output in the examples, from '42' to '52'.

Line 423 - Fix the semantic error 'the an' in the sentence.

Line 435 - Correct the class name in the text to match the class name in
the example.

Line 633 - A quotation mark was missing in the string that defines the
route '/' in router.get().